// src/common/enums/botTexts.ts
export enum BOT_TEXTS {
  WELCOME = "Добро пожаловать!",
  DEFAULT = "Получил сообщение",
  UNAVAILABLE = "Сервис не доступен :-(",
  RELOAD = "Попробуйте перезагрузить",
  TERMS_AGREEMENT = "Перед началом работы с ботом необходимо принять условия использования:",
  SHOW_COMMAND = "Доступные команды:",
}