// src/conversations/enums/conversationTermsTexts.ts
export enum CONVERSATION_TERMS_TEXTS {
  YES = "Принимаю",
  NO = "Не принимаю",
  YES_REPLY = "Здорово 😉, можете продолжить работу с ботом!",
  NO_REPLY = "К сожалению вы не можете продолжить работу с ботом 😟!",
  DESCRIPTION = `Список условий и отказ от ответственности.\r\nВы используете этого бота на свой страх и риск.\r\nПсихологический центр "DOM" не несет никакой ответственности.`,
}